Fix -Wsign-conversion warnings.

+++ src/daemon/common.c
@@ -257,8 +257,8 @@ ssize_t sread (int fd, void *buf, size_t count)
 
 		assert ((0 > status) || (nleft >= (size_t)status));
 
-		nleft = nleft - status;
-		ptr   = ptr   + status;
+		nleft = nleft - ((size_t) status);
+		ptr   = ptr   + ((size_t) status);
 	}
 
 	return (0);
@@ -284,8 +284,8 @@ ssize_t swrite (int fd, const void *buf, size_t count)
 		if (status < 0)
 			return (status);
 
-		nleft = nleft - status;
-		ptr   = ptr   + status;
+		nleft = nleft - ((size_t) status);
+		ptr   = ptr   + ((size_t) status);
 	}
 
 	return (0);
@@ -665,8 +665,8 @@ int check_create_dir (const char *file_orig)
 		 * Join the components together again
 		 */
 		dir[0] = '/';
-		if (strjoin (dir + path_is_absolute, dir_len - path_is_absolute,
-					fields, i + 1, "/") < 0)
+		if (strjoin (dir + path_is_absolute, (size_t) (dir_len - path_is_absolute),
+					fields, (size_t) (i + 1), "/") < 0)
 		{
 			ERROR ("strjoin failed: `%s', component #%i", file_orig, i);
 			return (-1);
